# docs/swagger.yaml
definitions:
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  models.ReputationScore:
    properties:
      attendanceRate:
        description: 出席率 (0-100)
        type: number
      behaviorRating:
        description: 行為評分 (1-5)
        type: number
      cancelledMatches:
        type: integer
      completedMatches:
        type: integer
      id:
        type: string
      overallScore:
        description: 綜合分數 (0-100)
        type: number
      punctualityScore:
        description: 準時度 (0-100)
        type: number
      skillAccuracy:
        description: 技術等級準確度 (0-100)
        type: number
      totalMatches:
        type: integer
      updatedAt:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 關聯
      userId:
        type: string
    type: object
...
  usecases.ReputationStats:
    properties:
      activeUsers:
        type: integer
      averageScore:
        type: number
      highReputationUsers:
        type: integer
      totalUsers:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 網球平台後端 API 服務
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 網球平台 API
  version: "1.0"
paths:
  /api/reputation/leaderboard:
    get:
      consumes:
      - application/json
      description: 獲取信譽分數排行榜
      parameters:
      - default: 50
        description: 返回數量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ReputationScore'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 獲取信譽排行榜
      tags:
      - reputation
  /api/reputation/stats:
    get:
      consumes:
      - application/json
      description: 獲取平台信譽系統的統計信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecases.ReputationStats'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 獲取信譽統計信息
      tags:
      - reputation
  /api/reputation/users/{userId}/attendance:
    post:
      consumes:
      - application/json
      description: 記錄用戶的比賽出席情況，影響出席率評分
      parameters:
      - description: 用戶ID
        in: path
        name: userId
        required: true
        type: string
      - description: 出席記錄請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RecordMatchAttendanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 記錄比賽出席情況
      tags:
      - reputation
  /api/reputation/users/{userId}/behavior-review:
    post:
      consumes:
      - application/json
      description: 對其他用戶提交行為評價，影響其行為評分
      parameters:
      - description: 被評價用戶ID
        in: path
        name: userId
        required: true
        type: string
      - description: 行為評價請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubmitBehaviorReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 提交行為評價
      tags:
      - reputation
  /api/reputation/users/{userId}/history:
    get:
      consumes:
      - application/json
      description: 獲取指定用戶的詳細信譽歷史記錄
      parameters:
      - description: 用戶ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReputationHistory'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 獲取用戶信譽歷史記錄
      tags:
      - reputation
  /api/reputation/users/{userId}/punctuality:
    post:
      consumes:
      - application/json
      description: 記錄用戶的比賽準時情況，影響準時度評分
      parameters:
      - description: 用戶ID
        in: path
        name: userId
        required: true
        type: string
      - description: 準時記錄請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RecordMatchPunctualityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 記錄比賽準時情況
      tags:
      - reputation
  /api/reputation/users/{userId}/score:
    get:
      consumes:
      - application/json
      description: 獲取指定用戶的信譽分數和統計信息
      parameters:
      - description: 用戶ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReputationScore'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 獲取用戶信譽分數
      tags:
      - reputation
  /api/reputation/users/{userId}/skill-accuracy:
    post:
      consumes:
      - application/json
      description: 記錄用戶的技術等級準確度，影響技術準確度評分
      parameters:
      - description: 用戶ID
        in: path
        name: userId
        required: true
        type: string
      - description: 技術準確度記錄請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RecordSkillLevelAccuracyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 記錄技術等級準確度
      tags:
      - reputation
  /api/reputation/users/{userId}/update-ntrp:
    post:
      consumes:
      - application/json
      description: 基於信譽系統數據自動調整用戶NTRP等級
      parameters:
      - description: 用戶ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 更新用戶NTRP等級
      tags:
      - reputation
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: 發送密碼重設郵件
      parameters:
      - description: 忘記密碼請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 忘記密碼
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 用戶登入系統
      parameters:
      - description: 登入請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: 用戶登入
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: 用戶登出系統，撤銷刷新令牌
      parameters:
      - description: 登出請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 用戶登出
      tags:
      - auth
  /api/v1/auth/oauth/{provider}:
    get:
      description: 獲取指定提供商的 OAuth 授權 URL
      parameters:
      - description: OAuth 提供商
        enum:
        - google
        - facebook
        - apple
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 獲取 OAuth 授權 URL
      tags:
      - auth
  /api/v1/auth/oauth/{provider}/callback:
    post:
      consumes:
      - application/json
      description: 處理 OAuth 提供商的回調並完成登入
      parameters:
      - description: OAuth 提供商
        enum:
        - google
        - facebook
        - apple
        in: path
        name: provider
        required: true
        type: string
      - description: OAuth 登入請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OAuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: OAuth 回調處理
      tags:
      - auth
  /api/v1/auth/oauth/{provider}/link:
    post:
      consumes:
      - application/json
      description: 將 OAuth 帳號關聯到當前用戶
      parameters:
      - description: OAuth 提供商
        enum:
        - google
        - facebook
        - apple
        in: path
        name: provider
        required: true
        type: string
      - description: 關聯 OAuth 帳號請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LinkOAuthAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 關聯 OAuth 帳號
      tags:
      - auth
  /api/v1/auth/oauth/{provider}/unlink:
    delete:
      consumes:
      - application/json
      description: 解除當前用戶與指定 OAuth 提供商的關聯
      parameters:
      - description: OAuth 提供商
        enum:
        - google
        - facebook
        - apple
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 解除關聯 OAuth 帳號
      tags:
      - auth
  /api/v1/auth/oauth/accounts:
    get:
      description: 獲取當前用戶已關聯的所有 OAuth 帳號列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 獲取已關聯的 OAuth 帳號
      tags:
      - auth
...
  /api/v1/chat/ws:
    get:
      description: 建立 WebSocket 連接用於即時聊天
      responses: {}
      security:
      - BearerAuth: []
      summary: WebSocket 連接
      tags:
      - chat
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"